Fix up plugin and tool priority code:
  * Priority is now a double instead of an int, similar to the
    MenuEntry's weight parameter. This makes it essentially always
    possible to squeeze a new processor or tool between two others in
    terms of priority.
  * Two new priority presets ("very high" and "very low") now exist.
  * The priorities of several processor plugins and tools have been
    adjusted.
  * The "first" and "last" priorities are now really always first and
    last, through the use of negative and positive infinity
    respectively. It is no longer possible to insert a processor before
    one with first priority, or after one with last priority.
  * The javadoc for several priority-related elements has been updated.

git-svn-id: http://code.imagej.net/svn/imagej/trunk@4401 be23795f-a60e-40d4-ba19-055798872e57

+++ core/ext/src/main/java/imagej/ext/plugin/DynamicPluginInfo.java
@@ -88,7 +88,7 @@ public class DynamicPluginInfo extends DefaultModuleInfo {
 	}
 
 	@Override
-	public int getPriority() {
+	public double getPriority() {
 		return info.getPriority();
 	}
 
@@ -123,7 +123,7 @@ public class DynamicPluginInfo extends DefaultModuleInfo {
 	}
 
 	@Override
-	public void setPriority(final int priority) {
+	public void setPriority(final double priority) {
 		info.setPriority(priority);
 	}
